+++ src/app/(main)/wrongbook/page.tsx
@@ -85,6 +85,7 @@ export default function WrongbookPage() {
   }, [page]);
 
   useEffect(() => {
+    // eslint-disable-next-line react-hooks/set-state-in-effect
     load();
   }, [load]);